상품수정

+++ page/screen/mypage/SettingWithdrawal.tsx
@@ -6,10 +6,10 @@
  * @flow strict-local
  */
 
-import React,{useState} from 'react';
+import React, { useState } from 'react';
 import {
-    Alert,
-  SafeAreaView, ScrollView, Text, View,StyleSheet, FlatList, Image,Button, TouchableOpacity
+  Alert,
+  SafeAreaView, ScrollView, Text, View, StyleSheet, FlatList, Image, Button, TouchableOpacity
 } from 'react-native';
 import { useNavigation } from '@react-navigation/native';
 import { StackNavigationProp } from '@react-navigation/stack';
@@ -32,35 +32,39 @@ import * as UserInfoAction from '../../../redux/actions/UserInfoAction'
 
 
 const SettingWithdrawal = () => {
-  const {t} = useTranslation()  
+  const { t } = useTranslation()
   const navigation = useNavigation<StackNavigationProp<MainNavigatorParams>>();
-  const userInfo = useSelector((state:any) => state.userInfo);
-  const myLocation = useSelector((state:any) => state.myLocation);
+  const userInfo = useSelector((state: any) => state.userInfo);
+  const myLocation = useSelector((state: any) => state.myLocation);
   const dispatch = useDispatch()
   /** redux 상태관리 */
 
-  const [withdrawal,setWithdrawal] = useState(false)
+  const [withdrawal, setWithdrawal] = useState(false)
 
-  const accccction = ()=>{
-    if(withdrawal==true){
-      Alert.alert(t('탈퇴하시겠습니까?'),''
-      ,
-      [
-        {text: '탈퇴', onPress: async() => {
-          await AsyncStorage.removeItem('userIdx')
-          DeleteUser();
-        }, style:'cancel'},
-        {
-          text: '취소',
-          onPress: () => {
-            
+  console.log('userInfo', userInfo);
+
+  const accccction = () => {
+    if (withdrawal == true) {
+      Alert.alert(t('탈퇴하시겠습니까?'), ''
+        ,
+        [
+          {
+            text: '탈퇴', onPress: async () => {
+              await AsyncStorage.removeItem('userIdx')
+              DeleteUser();
+            }, style: 'cancel'
           },
-          style: 'destructive',
-        },
-      ]
+          {
+            text: '취소',
+            onPress: () => {
+
+            },
+            style: 'destructive',
+          },
+        ]
       )
-    } else{
-      Alert.alert('약관에 동의해주세요','')
+    } else {
+      Alert.alert('약관에 동의해주세요', '')
     }
   }
 
@@ -68,54 +72,54 @@ const SettingWithdrawal = () => {
     await client<any>({
       method: 'get',
       url: '/user/memout',
-      params:{
-        mt_idx:userInfo.idx
+      params: {
+        mt_idx: userInfo.idx
       }
-      }).then(
-        res=>{
-          console.log(res.data)
-          dispatch(UserInfoAction.logOut());
-          navigation.reset({routes:[{name:'SelectLogin'}]});
-        }
-      ).catch(
-        err=>console.log(err)
-      )
-    };
+    }).then(
+      res => {
+        console.log(res.data)
+        dispatch(UserInfoAction.logOut());
+        navigation.reset({ routes: [{ name: 'SelectLogin' }] });
+      }
+    ).catch(
+      err => console.log(err)
+    )
+  };
 
-    return (
-        <SafeAreaView style={{flex:1,backgroundColor:'#fff'}}>
-          <BackHeader title={t('탈퇴하기')}/>
-            <ScrollView style={{paddingHorizontal:20}}>
-              <Text style={[style.text_b,{fontSize:22,color:colors.BLACK_COLOR_1,paddingTop:30,}]}>
-                {userInfo.mt_nickname} {t('님')}
-              </Text>
-              <Text style={[style.text_b,{fontSize:22,color:colors.BLACK_COLOR_1}]}>
-              {t('정말로 탈퇴하시겠습니까?')}
-              </Text>
-              <Text style={[style.text_re,{fontSize:15,color:colors.BLACK_COLOR_1,marginRight:20,paddingVertical:25,lineHeight:23}]}>
-              {t('계정을 삭제하면 판매내역, 평가받은 만족도와 후기, 관심, 채팅 모든 활동 정보가 삭제됩니다. 계정 삭제후 7일간 다시 가입할 수 없습니다.')}
-              </Text>
+  return (
+    <SafeAreaView style={{ flex: 1, backgroundColor: '#fff' }}>
+      <BackHeader title={t('탈퇴하기')} />
+      <ScrollView style={{ paddingHorizontal: 20 }}>
+        <Text style={[style.text_b, { fontSize: 22, color: colors.BLACK_COLOR_1, paddingTop: 30, }]}>
+          {userInfo.mt_nickname} {t('님')}
+        </Text>
+        <Text style={[style.text_b, { fontSize: 22, color: colors.BLACK_COLOR_1 }]}>
+          {t('정말로 탈퇴하시겠습니까?')}
+        </Text>
+        <Text style={[style.text_re, { fontSize: 15, color: colors.BLACK_COLOR_1, marginRight: 20, paddingVertical: 25, lineHeight: 23 }]}>
+          {t('계정을 삭제하면 판매내역, 평가받은 만족도와 후기, 관심, 채팅 모든 활동 정보가 삭제됩니다. 계정 삭제후 7일간 다시 가입할 수 없습니다.')}
+        </Text>
 
-              <TouchableOpacity style={{flexDirection:'row'}} onPress={()=>setWithdrawal(!withdrawal)}>
-                {!withdrawal? 
-                <Image style={{width:22,height:22}} source={require('../../../assets/img/check_off.png')}/> :
-                <Image style={{width:22,height:22}} source={require('../../../assets/img/check_on.png')}/>
-                }
-                <Text style={[style.text_me,{fontSize:15,color:colors.BLACK_COLOR_1,marginLeft:10}]}>
-                {t('네, 확인했습니다.')}
-                </Text>
-              </TouchableOpacity>
-            </ScrollView>
-            <TouchableOpacity 
-                onPress = {accccction}
-                style={[{backgroundColor:withdrawal? colors.GREEN_COLOR_2:colors.GRAY_COLOR_2,alignItems:'center',justifyContent:'center', height:60}]}>
-                <Text style={[style.text_sb,{color:colors.WHITE_COLOR, fontSize:18}]}> 
-                {t('탈퇴하기')}
-                </Text>
-            </TouchableOpacity>
-            <BackHandlerCom />
-        </SafeAreaView>
-    );
+        <TouchableOpacity style={{ flexDirection: 'row' }} onPress={() => setWithdrawal(!withdrawal)}>
+          {!withdrawal ?
+            <Image style={{ width: 22, height: 22 }} source={require('../../../assets/img/check_off.png')} /> :
+            <Image style={{ width: 22, height: 22 }} source={require('../../../assets/img/check_on.png')} />
+          }
+          <Text style={[style.text_me, { fontSize: 15, color: colors.BLACK_COLOR_1, marginLeft: 10 }]}>
+            {t('네, 확인했습니다.')}
+          </Text>
+        </TouchableOpacity>
+      </ScrollView>
+      <TouchableOpacity
+        onPress={accccction}
+        style={[{ backgroundColor: withdrawal ? colors.GREEN_COLOR_2 : colors.GRAY_COLOR_2, alignItems: 'center', justifyContent: 'center', height: 60 }]}>
+        <Text style={[style.text_sb, { color: colors.WHITE_COLOR, fontSize: 18 }]}>
+          {t('탈퇴하기')}
+        </Text>
+      </TouchableOpacity>
+      <BackHandlerCom />
+    </SafeAreaView>
+  );
 };
 
 export default SettingWithdrawal;
